+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Attributes/AmbientCoefficient.cpp
@@ -32,8 +32,8 @@ AmbientCoefficient::AmbientCoefficient()
 {
 }
 
-void	AmbientCoefficient::Process( uint32_t inStartOffset,
-									uint32_t inEndOffset )
+void	AmbientCoefficient::Process( size_t inStartOffset,
+									size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 4)
 	{
