+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Attributes/AmbientCoefficient.h
@@ -30,6 +30,6 @@ class AmbientCoefficient : public TypeHandler
 public:
 							AmbientCoefficient();
 	
-	virtual void			Process( uint32_t inStartOffset,
-									uint32_t inEndOffset );
+	virtual void			Process( size_t inStartOffset,
+									size_t inEndOffset );
 };
