+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Attributes/DiffuseColor.cpp
@@ -32,8 +32,8 @@ DiffuseColor::DiffuseColor()
 {
 }
 
-void	DiffuseColor::Process( uint32_t inStartOffset,
-									uint32_t inEndOffset )
+void	DiffuseColor::Process( size_t inStartOffset,
+									size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 12)
 	{
