+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Attributes/SpecularColor.cpp
@@ -33,8 +33,8 @@ SpecularColor::SpecularColor()
 {
 }
 
-void	SpecularColor::Process( uint32_t inStartOffset,
-									uint32_t inEndOffset )
+void	SpecularColor::Process( size_t inStartOffset,
+									size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 12)
 	{
