+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Cameras/OrthographicCamera.cpp
@@ -34,8 +34,8 @@ OrthographicCamera::OrthographicCamera()
 }
 
 
-void	OrthographicCamera::Process( uint32_t inStartOffset,
-										uint32_t inEndOffset )
+void	OrthographicCamera::Process( size_t inStartOffset,
+										size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 16)
 	{
