+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Cameras/ViewPlaneCamera.cpp
@@ -33,8 +33,8 @@ ViewPlaneCamera::ViewPlaneCamera()
 }
 
 
-void	ViewPlaneCamera::Process( uint32_t inStartOffset,
-										uint32_t inEndOffset )
+void	ViewPlaneCamera::Process( size_t inStartOffset,
+										size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 20)
 	{
