+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Groups/DisplayGroupState.cpp
@@ -49,10 +49,10 @@ DisplayGroupState::DisplayGroupState()
 {
 }
 
-void	DisplayGroupState::Process( uint32_t inStartOffset,
-										uint32_t inEndOffset )
+void	DisplayGroupState::Process( size_t inStartOffset,
+										size_t inEndOffset )
 {
-	uint32_t dataLen = inEndOffset - inStartOffset;
+	size_t dataLen = inEndOffset - inStartOffset;
 	
 	if (dataLen != 4)
 	{
