+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Styles/FillStyle.h
@@ -30,6 +30,6 @@ class FillStyle : public TypeHandler
 public:
 							FillStyle();
 	
-	virtual void			Process( uint32_t inStartOffset,
-									uint32_t inEndOffset );
+	virtual void			Process( size_t inStartOffset,
+									size_t inEndOffset );
 };
