+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Styles/InterpolationStyle.cpp
@@ -35,8 +35,8 @@ InterpolationStyle::InterpolationStyle()
 {
 }
 
-void	InterpolationStyle::Process( uint32_t inStartOffset,
-									uint32_t inEndOffset )
+void	InterpolationStyle::Process( size_t inStartOffset,
+									size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 4)
 	{
