+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Styles/OrientationStyle.cpp
@@ -42,8 +42,8 @@ OrientationStyle::OrientationStyle()
 {
 }
 
-void	OrientationStyle::Process( uint32_t inStartOffset,
-									uint32_t inEndOffset )
+void	OrientationStyle::Process( size_t inStartOffset,
+									size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 4)
 	{
