+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Styles/ReceiveShadowsStyle.cpp
@@ -32,8 +32,8 @@ ReceiveShadowsStyle::ReceiveShadowsStyle()
 {
 }
 
-void	ReceiveShadowsStyle::Process( uint32_t inStartOffset,
-									uint32_t inEndOffset )
+void	ReceiveShadowsStyle::Process( size_t inStartOffset,
+									size_t inEndOffset )
 {
 	if (inEndOffset != inStartOffset + 4)
 	{
