+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Styles/SubdivisionStyle.cpp
@@ -37,8 +37,8 @@ SubdivisionStyle::SubdivisionStyle()
 {
 }
 
-void	SubdivisionStyle::Process( uint32_t inStartOffset,
-										uint32_t inEndOffset )
+void	SubdivisionStyle::Process( size_t inStartOffset,
+										size_t inEndOffset )
 {
 	if (inEndOffset - inStartOffset < 8)
 	{
