+++ SDK/Extras/Textify3DMF/Source/TypeHandlers/Transforms/Quaternion.cpp
@@ -32,10 +32,10 @@ Quaternion::Quaternion()
 {
 }
 
-void	Quaternion::Process( uint32_t inStartOffset,
-										uint32_t inEndOffset )
+void	Quaternion::Process( size_t inStartOffset,
+										size_t inEndOffset )
 {
-	uint32_t dataLen = inEndOffset - inStartOffset;
+	size_t dataLen = inEndOffset - inStartOffset;
 
 	if (dataLen != 16)
 	{
